refactor everything
