+++ dashboard/src/main/home/infrastructure-dashboard/ClusterSaveButton.tsx
@@ -2,28 +2,26 @@ import React from "react";
 
 import Button from "components/porter/Button";
 
-import { type ButtonStatus } from "./ClusterTabs";
+import { useClusterContext } from "./ClusterContextProvider";
+import { useClusterFormContext } from "./ClusterFormContextProvider";
 
 type Props = {
-  status: ButtonStatus;
-  isDisabled: boolean;
   height?: string;
   disabledTooltipPosition?: "top" | "bottom" | "left" | "right";
-  loadingText?: string;
 };
 const ClusterSaveButton: React.FC<Props> = ({
-  status,
-  isDisabled,
   height,
   disabledTooltipPosition,
-  loadingText,
 }) => {
+  const { updateClusterButtonProps } = useClusterFormContext();
+  const { isClusterUpdating } = useClusterContext();
+
   return (
     <Button
       type="submit"
-      status={status}
-      loadingText={loadingText}
-      disabled={isDisabled}
+      status={updateClusterButtonProps.status}
+      loadingText={updateClusterButtonProps.loadingText}
+      disabled={updateClusterButtonProps.isDisabled || isClusterUpdating}
       disabledTooltipMessage={
         "Please wait for the current update to complete before updating again."
       }
