+++ dashboard/src/main/home/infrastructure-dashboard/forms/gcp/ConfigureGKECluster.tsx
@@ -11,22 +11,15 @@ import VerticalSteps from "components/porter/VerticalSteps";
 import { CloudProviderGCP } from "lib/clusters/constants";
 import { type ClientClusterContract } from "lib/clusters/types";
 
+import { useClusterFormContext } from "../../ClusterFormContextProvider";
 import NodeGroups from "../../shared/NodeGroups";
 import { BackButton, Img } from "../CreateClusterForm";
 
 type Props = {
   goBack: () => void;
-  createButtonProps: {
-    status: "loading" | JSX.Element | "success" | "";
-    isDisabled: boolean;
-    loadingText: string;
-  };
 };
 
-const ConfigureGKECluster: React.FC<Props> = ({
-  goBack,
-  createButtonProps,
-}) => {
+const ConfigureGKECluster: React.FC<Props> = ({ goBack }) => {
   const [currentStep, _setCurrentStep] = useState<number>(4);
 
   const {
@@ -34,6 +27,8 @@ const ConfigureGKECluster: React.FC<Props> = ({
     register,
     formState: { errors },
   } = useFormContext<ClientClusterContract>();
+
+  const { updateClusterButtonProps } = useClusterFormContext();
 
   return (
     <div>
@@ -118,9 +113,9 @@ const ConfigureGKECluster: React.FC<Props> = ({
             <Spacer y={0.5} />
             <Button
               type="submit"
-              status={createButtonProps.status}
-              disabled={createButtonProps.isDisabled}
-              loadingText={createButtonProps.loadingText}
+              status={updateClusterButtonProps.status}
+              disabled={updateClusterButtonProps.isDisabled}
+              loadingText={updateClusterButtonProps.loadingText}
             >
               Submit
             </Button>
